Add typehints and typechecking with mypy
This commit adds typehints to all functions, and some variables,
and adds a call to mypy in the scripts/test script.

Typechecking the calls to repo.get_contents() was tricky, as
this method can also return a list of files if called on a
dictionary. Moreover, repo.get_contents() was being called
once to check whether a file existed, and again to get the
contents of that file. Refactoring the code has eliminated
that extra API call.