Describe the relationship between standard docs and org defaults
Previously the script logged whether repos had a CONTRIBUTING and
a CODE_OF_CONDUCT, and whether that document was the same (via
str equals) as the organisation default. In this commit, we
make this relationship more specific, by noting:

* Whether the organisation does not have a default file.
* Whether the file in the repository is missing.
* Whether the file is identical to the org default.
* Whether the file contains a link to the org default.
* Whether the file differs (and does not link to) the org default.

This commit also fixes a bug where the contents of files was being
returns in an encoded format. This is fixed with:

    content_file.decoded_content.decode()