Skip archived repos
Repositories that are archived are not expected to be maintained,
and therefore may not have the current set of health checks and
settings. The information we could gather about arhvied repos
is not useful, and it's easier not to collect it. This makes
the script much quicker when dealing with large organisations,
and makes the output spreadsheet a little less unweildy.